+++ test/ir_test.cpp
@@ -28,9 +28,9 @@ TEST(IRTest, BasicAssertions) {
     //      VarType::LoopInductionVariable), std::make_pair(0,
     //      VarType::LoopInductionVariable)});
 
-    // std::vector<Int> coef_memory({1, 2, 3, 5, 7, 11, 13, 17});
+    // std::vector<int64_t> coef_memory({1, 2, 3, 5, 7, 11, 13, 17});
     // std::vector<size_t> coef_offsets({0, 1, 2, 3, 5, 8});
-    // VoV<Int> coef = VoV<Int>(toVector(coef_memory), toVector(coef_offsets));
+    // VoV<int64_t> coef = VoV<int64_t>(toVector(coef_memory), toVector(coef_offsets));
     // llvm::SmallVector<size_t> pvc_memory({0, 0, 1, 0, 0, 2, 0, 1, 2});
     // std::vector<size_t> innerOffsets({0, 0, 0, 1, 0, 2, 0, 0, 1, 0, 0, 2,
     // 5}); printf("innOff len: %d\n", innerOffsets.size()); std::vector<size_t>
@@ -43,32 +43,32 @@ TEST(IRTest, BasicAssertions) {
     //                   toVector(outerOffsets), memBuffer);
 
     llvm::SmallVector<
-        std::pair<Polynomial::Multivariate<intptr_t, Polynomial::Monomial>,
+        std::pair<Polynomial::Multivariate<int64_t, Polynomial::Monomial>,
                   VarID>,
         2>
         inds;
     inds.emplace_back(
-        Polynomial::Multivariate<intptr_t, Polynomial::Monomial>(1),
+        Polynomial::Multivariate<int64_t, Polynomial::Monomial>(1),
         VarID(2, VarType::LoopInductionVariable));
     inds.emplace_back(
-        Polynomial::MultivariateTerm<intptr_t, Polynomial::Monomial>(
+        Polynomial::MultivariateTerm<int64_t, Polynomial::Monomial>(
             2, Polynomial::Monomial(Polynomial::ID{0})),
         VarID(8, VarType::Memory));
     inds.emplace_back(
-        Polynomial::MultivariateTerm<intptr_t, Polynomial::Monomial>(
+        Polynomial::MultivariateTerm<int64_t, Polynomial::Monomial>(
             3, Polynomial::Monomial(Polynomial::ID{0}, Polynomial::ID{1})),
         VarID(18, VarType::Term));
-    Polynomial::Multivariate<intptr_t, Polynomial::Monomial> p3(5);
-    p3.addTerm(Polynomial::MultivariateTerm<intptr_t, Polynomial::Monomial>(
+    Polynomial::Multivariate<int64_t, Polynomial::Monomial> p3(5);
+    p3.addTerm(Polynomial::MultivariateTerm<int64_t, Polynomial::Monomial>(
         7, Polynomial::Monomial(Polynomial::ID{0})));
     inds.emplace_back(p3, VarID(3, VarType::LoopInductionVariable));
-    Polynomial::Multivariate<intptr_t, Polynomial::Monomial> p4(11);
-    p4.addTerm(Polynomial::MultivariateTerm<intptr_t, Polynomial::Monomial>(
+    Polynomial::Multivariate<int64_t, Polynomial::Monomial> p4(11);
+    p4.addTerm(Polynomial::MultivariateTerm<int64_t, Polynomial::Monomial>(
         13, Polynomial::Monomial(Polynomial::ID{0}, Polynomial::ID{3})));
-    p4.addTerm(Polynomial::MultivariateTerm<intptr_t, Polynomial::Monomial>(
+    p4.addTerm(Polynomial::MultivariateTerm<int64_t, Polynomial::Monomial>(
         17, Polynomial::Monomial(Polynomial::ID{0}, Polynomial::ID{1},
                                  Polynomial::ID{2})));
-    p4.addTerm(Polynomial::MultivariateTerm<intptr_t, Polynomial::Monomial>(
+    p4.addTerm(Polynomial::MultivariateTerm<int64_t, Polynomial::Monomial>(
         11, Polynomial::Monomial(Polynomial::ID{0}, Polynomial::ID{0},
                                  Polynomial::ID{2})));
     inds.emplace_back(p4, VarID(0, VarType::LoopInductionVariable));
